requirements.txt
python-dotenv==1.0.1
xarray==2024.7.0
rioxarray==0.18.1
cftime==1.6.4.post1
dask[dataframe]==2024.8.0
Cartopy==0.23.0
cdsapi==0.7.2
h5netcdf==1.3.0
matplotlib==3.9.2
seaborn==0.13.2
scipy==1.15.0
plotly==5.24.1
Pygments==2.18.0
osmnx==2.0.1
scikit-learn==1.5.2
folium==0.18.0
geopy==2.4.1
nbformat>=4.2.0
rasterio==1.4.2
pandas_access==0.0.1
sqlalchemy==2.0.36
OWSLib==0.31.0
shap==0.46.0
lightgbm==4.5.0
mapie==0.9.1
quantile-forest==1.3.11
tensorflow==2.18.0
papermill==2.6.0
ipywidgets==8.1.5
interpret==0.6.9
tabulate==0.9.0